// desenvolvimento/hibrido/src/pages/shared/model/firebase-config.model.ts
export interface IFirebaseConfig {
    agenda: string;
    avaliacao: string;
    avaliacao_usuario: string;
    caracteristica: string;
    caracteristica_tipo_ponto_interesse: string;
    item: string;
    notificacao: string;
    ponto_interesse: string;
    preferencia_usuario: string;
    rota: string;
    tipo_agenda: string;
    tipo_dado: string;
    tipo_ponto_interesse: string;
    tipo_transporte: string;
    tipo: string;
    transporte: string;
    usuario: string;
}